Payment_BE

+++ PRN231_2_EventFlowerExchange_BE/BusinessObject/DTO/Request/CreateFlowerDTO.cs
@@ -12,7 +12,7 @@ namespace BusinessObject.DTO.Request
         public string Type { get; set; }
         public string Image { get; set; }
         public string Description { get; set; }
-        public decimal PricePerUnit { get; set; }
+        public double PricePerUnit { get; set; }
         public string Origin { get; set; }
         public string Color { get; set; }
         public int RemainingQuantity { get; set; }
